pyproject.toml
[project]
name = "openai-tts-test"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Mizuki", email = "mbaba@kanolab.net" }
]
dependencies = [
    "openai>=1.1.2",
    "python-dotenv>=1.0.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true
